Separate row components from scene scrape dialog
